Improve TxOperations API
Improve TxOperations API
