+++ packages/presentation/src/attributes.ts
@@ -8,17 +8,14 @@ export interface KeyedAttribute {
   attr: AnyAttribute
 }
 
-export async function updateAttribute (client: Client & TxOperations, object: Doc, _class: Ref<Class<Doc>>, attribute: KeyedAttribute, value: any): Promise<void> {
+export async function updateAttribute (client: TxOperations, object: Doc, _class: Ref<Class<Doc>>, attribute: KeyedAttribute, value: any): Promise<void> {
   const doc = object
   const attributeKey = attribute.key
   const attr = attribute.attr
   if (client.getHierarchy().isMixin(attr.attributeOf)) {
     await client.updateMixin(doc._id, _class, doc.space, attr.attributeOf, { [attributeKey]: value })
-  } else if (client.getHierarchy().isDerived(object._class, core.class.AttachedDoc)) {
-    const adoc = object as AttachedDoc
-    await client.updateCollection(_class, object.space, adoc._id, adoc.attachedTo, adoc.attachedToClass, adoc.collection, { [attributeKey]: value })
   } else {
-    await client.updateDoc(_class, doc.space, doc._id, { [attributeKey]: value })
+    await client.update(object, { [attributeKey]: value })
   }
 }
